drop stdint refinements of cmake_example_min, just use standard int

+++ cmake_example_min/blink_lib.cpp
@@ -1,12 +1,12 @@
 #include "Arduino.h"
 #include "blink_lib.h"
 
-void blink_setup(uint8_t pin) {
+void blink_setup(int pin) {
   pinMode(pin, OUTPUT);
 }
 
 
-void blink(unsigned long duration, uint8_t pin) {
+void blink(long duration, int pin) {
   digitalWrite(pin, HIGH);   // set the LED on
   delay(duration);           // wait for a second
   digitalWrite(pin, LOW);    // set the LED off
